+++ Esp.Net.Tests/Esp.Net.Tests.csproj
@@ -53,7 +53,7 @@
     <Compile Include="Model\EspDisposableTests.cs" />
     <Compile Include="Stubs\StubEventSubject.cs" />
     <Compile Include="Stubs\StubRouter.cs" />
-    <Compile Include="Concurrency\WorkflowTests.cs" />
+    <Compile Include="Workflow\WorkflowTests.cs" />
     <Compile Include="Stubs\StubSubject.cs" />
     <Compile Include="HeldEvents\HeldEventTests.cs" />
     <Compile Include="Reactive\ReactiveTests.cs" />
