+++ Esp.Net/Esp.Net.csproj
@@ -44,7 +44,7 @@
     <Compile Include="..\AssemblyInfoCommon.cs">
       <Link>Properties\AssemblyInfoCommon.cs</Link>
     </Compile>
-    <Compile Include="Concurrency\AyncResultsEvent.cs" />
+    <Compile Include="Workflow\AyncResultsEvent.cs" />
     <Compile Include="EventContext.cs" />
     <Compile Include="EventProcessors.cs" />
     <Compile Include="Guard.cs" />
@@ -60,8 +60,8 @@
     <Compile Include="Model\DisposableBase.cs" />
     <Compile Include="Model\EspSerialDisposable.cs" />
     <Compile Include="ObservationStage.cs" />
-    <Compile Include="Concurrency\Workflow.cs" />
-    <Compile Include="Concurrency\Workflow.Steps.cs" />
+    <Compile Include="Workflow\Workflow.cs" />
+    <Compile Include="Workflow\Workflow.Steps.cs" />
     <Compile Include="Reactive\ModelObservable.cs" />
     <Compile Include="Reactive\ModelObserver.cs" />
     <Compile Include="Reactive\ModelSubject.cs" />
