+++ Esp.Net/HeldEvents/IEventDescription.cs
@@ -1,6 +1,6 @@
-﻿using System;
+﻿#if ESP_EXPERIMENTAL
+using System;
 
-#if ESP_EXPERIMENTAL
 namespace Esp.Net.HeldEvents
 {
     public interface IEventDescription
